--- src/bin/widgets.rs
@@ -1,1 +1,0 @@
-// Contains the reusable widgets that can be used in UI scenes
